requirements.txt
# WebSocket y OCPP
websockets==10.4
ocpp==0.8.1

# API
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.6

# DB y ORM
sqlalchemy==2.0.27
asyncpg==0.29.0      # si usas PostgreSQL
aiosqlite==0.19.0    # si usas SQLite

# Seguridad y JWT
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.6

# Fecha y parsing
python-dateutil==2.8.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Logging YAML
pyyaml==6.0.1

pydantic-settings>=0.1.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings>=0.1.3

# Seguridad y validación
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings>=0.1.3
email-validator==1.3.1

bcrypt>=4.0.0